back button
- Added new button in classification page that
allows the user to get to the classification page
of the image with image id one lower than the
currently displayed image.

- Changed implementation of skip button. The skip
button is now not part of the MultiButtonForm
class anymore, but implemented in plain html.

- The skip button now redirects the user to the
image with image id one larger than the current
image. Previously, it redirected the user to the
unlabelled image with highest image id.

- Removed function skip_clasification and
introduced function handle_move_button that
handles skip button and back button actions.