safer request handling for classification submit
- Fixed a bug, where an error would occur when the
user labels or skips the image with highest
DatasetID of a dataset. The user is now redirected
to the next unlabelled image with lowest ImageID
or, if no unlabelled images remain, he is
redirected to the classification summary page.

- Fixed a bug, where an error would occur if the
user clicks on the back button when being on the
classification page of the first image of a
dataset. In this case the user is now redirected
to the current page.