progress bar
- Added progress bar on the classification page,
that displays how many images in the current
dataset have already been labelled or skipped in
comparison to the total number of images in the
dataset both in absolute and relative numbers.

- Added convenience function
get_num_of_labelled_images to get the total number
of labelled and skipped images in a specified
dataset.