export labels
- Implemented the feature to export labeled
images for a specied dataset. The results can be
exported in csv or JSON format.

- Added new entry "export" on navbar that
navigates to new page, where the dataset to export
and the export format can be chosen.

- Added new convenience function
get_results_by_dataset that fetches image names
names and corresponding labels for a given
dataset.

- Added new file result_writers.py, where
write_to_csv and write_to_json functions are
defined.