add images to datasets
- Added entry "Modify datasets" to "Edit datasets"
ribbon of navbar.

- Added new endpoint
/config/select_datasets_to_edit that renders page
in which users can choose a dataset to add images
to or remove images from.

- Implemented adding images to dataset. Users can
choose another folder with images, that are
uploaded and added to an existing dataset.
Duplicate images (which already existed in the
datasest) are automatically discarded.

- Added new parameter start_dataset_index to
bulk_insert_images function, which shifts the
DatasetID values of inserted images by a constant
amount.